backend-car-app/requirements.txt
fastapi
uvicorn
prometheus-client
prometheus_fastapi_instrumentator
python-logging-loki
requests